pyproject.toml
[project]
name = "fastapi-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "cryptography>=45.0.5",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]
